added serve function
The function "serve" handles the serving of big files, read in 1k chunks,
before closing the connection at the end with a reply. Currently doesn't
handle multiple concurrent files, this would require storing the file cursor,
reopening and seeking at every "sent" event. For now i prefer to keep it
simple, and limit the download of concurrent files to < 512 bytes files or
to simply queue the download of bigger files with javascript.

Most big files are more conveniently stored in the cloud anyway (bootstrap
css, icons, fonts, js libraries and so on). Some files can be avoided altogether
like favicon.ico, some small files like the optional html5 manifest can be
useful in limiting the number of concurrent connections, a big cache time value
for static content could be sent in the proper header (for example all gz files
or all files if they never change), and most stuff can be included in the
index.html as datauri or embedded in script and style tags. Some files are only
loaded when needed by the browser, correctly placing images could help a lot.

So this doesn't sound to me like a huge limitation, but it could probably
be lifted away if needed.