Basic auth support for /edit path
This currently is the only line of code handling basic auth, it
could be easily removed or commented out.

a better way is to disable this check when pwd is nil, and possibly
to precompute the base64 value, it would use a bit more memory than
the clear pwd password but this would avoid a call to crypto.toBase64
at every /edit request.