a few commends implemented
actually download and edit can be combined into a single edit, because
javascript calls don't trigger a file download even with
the "Content-Disposition: attachment; filename=..." header
or with "applcation/octed-stream" Content-Type. Possibly one of these
headers could be removed too.
the list cmd is compact but requires cjson module, it could be
implemented by simple string concatenation, not sure with how much
memory.
the list cmd could also be moved to the browser side throug the
run cmd, like i hope to do with the delete and eventually compile
commands. I could give up on the preview command, or maybe not.