get Content-Length
Firefox and Chrome set this header correctly for multipart forms, so
use this value. It would have been trickier and more memory expensive
to check the end coundary of the multipart form because it can span
over 2 packets, requiring buffering of the tail of every packet.
But with accurate Content-Length this can be avoided.