The save function to handle big uploads
A single function to handle all 4 cases:
1) no multipart form data after the headers in the first packet (chrome)
2) multipart form data after the headers in the first packet (firefox)
3) multipart form data in the second packet (chrome)
4) binary data in every other packet up till the computed length (chrome and firefox)

The filename must be extracted from the multipart form data. It could be added
to the request url as an arg to a "upload" cmd or something like that, but i
have to parse the multipart form data anyway, so there is little point.

When it works it works very well, uploading files>=500kbytes presents
no problem at all. I've seen a few panics, possibly due to retransmission
attemps? i don't know. Of course don't even try to upload more than 1 file
at a time.