cosmetic in usage of wscb.
This code should also send back the output of wscb to the remote caller.
Sadly there is a chickend and egg problem with the bcast definition.
It may be possible that defining a local nil bcast before require("httpd")
and therefore before defining the wscb will do the trick, beed to test this
one again.
It could also be both wrapped and unwrapped by cjson.