remove title tag from head
i can never see anyway the title of my tabs because they are too small, but
most importantly this lets browsers to dinamically fill the title of the
window or tab with the name of the host, which is more important than
a fixed name when you are editing 10 different remote sensors.

Maybe an empty title tag would have the same effect and be easier to edit
for users.