added file size info to treeview
this is the trivial way to do it. The nicer way is to add a second
span to every li, put the size in its innerhtml alone, and make all
the last-child span below #tree float right and with background
inherit to remove the icon. sadly a fixed 10px padding-right is also
necessary to make them visible, not sure it is worth all the effort