Allows editing of gzipped text files
Since they are served with the correct Content-Encoding browsers
will treat them like any other text file. If saved they will be
overwritten by a not compressed version, keeping the .gz extension
This will not be a problem because the Content-Encoding is decided
based on the magic bytes, not on the extension.
It would be nice to compress them in the browser through javascript
and send the compressed version
there is a 5.9k gzip.min.js.gz that could do the trick:
https://github.com/imaya/zlib.js/blob/master/bin/gzip.min.js
there is a 3.6k rawdeflate.min.js.gz:
https://github.com/dankogai/js-deflate/blob/master/rawdeflate.js
that could be extended to cover the gzip format by adding crc32 as
already done by zlib.js
there is pako_deflate.min.js.gz at 8.5k also available from cdnsj:
https://github.com/nodeca/pako/blob/master/dist/pako_deflate.js
there is also an emscripten zopfli.gzip.min.js.gz port, but huge at 34k
https://github.com/imaya/zopfli.js/blob/master/bin/zopfli.gzip.min.js