Immediately reply to websocket connection with default return value of callback
This allows defining a callback that when called with nil argument returns the
current state of the sensors or something like that, so the client can readily
draw a graph or act on them without polling even once.

There could be a problem if the callback is not defined: should i return a pong?
or an empty string? or nothing at all?