Add a command prompt, a serial prompt and a restart button
The prompts are just an hack with a couple textareas, they would need
much more tweaks and testing on other browsers and platforms.

They also probably could be styled differently with float:left and
clear:both to avoid the fixed positioning, and/or their size and
positioning should be better handled with js.

Serial isn't collected to the backend