Fix basic auth in Firefox
The problem was that Firefox puts the Authorization field at the
bottom of the headers, so the pattern looking for the trailing \r
always failed because the headers were split from the body. Since
lua patterns don't have a construct for "or" the easiest fix is to
search directly for the base64 characters.