Nodemcu websocket implementation has a "s" in place of a "S"
http headers are case insensitive, but to keep the patterns as
short as possible i just match what firefox and chrome sends out,
which is exactly as written in the rfc, basically CamelCase.

The websocket implementation included in Nodemcu uses "Sec-Websocket-Key"
instead of the more common "Sec-WebSocket-Key". It is fully compliant,
but it would be better if it followed the rfc, firefox and chrome.

In this case i can just update the pattern to match both cases, it still
might be a good idea to send a pull request.